backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
clarifai==11.6.4
clarifai-grpc==11.6.2
clarifai-protocol==0.0.25
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-schema-viewer==0.5.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ijson==3.4.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
multidict==6.6.3
numpy==2.3.1
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
schema==0.7.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.12
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
daphne==4.1.2
